Cleanup UA Parser
